Fix threading issues on dialog boxes
On macOS and Windows, all GUI elements are expected to run on the main thread. The CaptureThread creates dialog boxes which run on a separate thread from the MainWindow class, thus creating issues on both platforms.

This commit's approach is to add signals in the CaptureThread class and slots in MainWindow, and move all GUI code out of the CaptureThread class.